LayoutTests/imported/w3c/web-platform-tests/payment-request/payment-request-abort-method.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.

PASS Throws if the promise [[state]] is not "interactive" 
FAIL Test for PaymentRequest.abort() method promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: test_driver"
FAIL Test for PaymentRequest.abort() method 1 promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: test_driver"
PASS Calling abort() multiple times is always a new object. 
